+++ src/components/ui-nav-list.vue
@@ -1,3 +1,4 @@
+/** -> "../components/def-nav.vue" */
 <template>
   <!-- nav内で利用するリスト一つ一つ -->
   <div>
